blender_extension/pyproject.toml
[tool.poetry]
name = "blender-extension"
version = "0.1.0"
description = "A simple Blender extension"
authors = ["Benjamin D"]
readme = "README.md"
packages = [
    { include = "*.py" }
]

[tool.poetry.dependencies]
python = "=3.11.9"
scipy = "^1.15.1"
fake-bpy-module = "^20250215"
torch = "^2.6.0"
matplotlib = "^3.10.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"